Refactor site build
