+++ docs/.werf/nginx.conf
@@ -183,7 +183,7 @@ http {
     server {
         charset utf-8;
         listen 80;
-        server_name ru.werf.io ru.werf.test.flant.com ru.werf.stage.flant.com ru.werfng.test.flant.com;
+        server_name ru.werf.io ru.ng.werf.io ru.werf.test.flant.com ru.werf.stage.flant.com ru.werfng.test.flant.com;
 
         root        /app/ru;
         index       index.html;
