--- docs/_includes/en/banner.html
@@ -1,1 +1,0 @@
-{%- include en/guides_banner.html %}
