--- docs/_includes/footer-common.html
@@ -1,24 +1,0 @@
-<div class="footer">
-    <div class="footer__credits">
-        {%- if page.lang == "ru" %}
-        Сделано с <span class="page__icon page__icon_heart"></span> в компании <a href="{{ site.company_url }}" target="_blank">{{ site.company_name }}</a>
-        {%- else %}
-        Made with <span class="page__icon page__icon_heart"></span> by <a href="{{ site.company_url }}" target="_blank">{{ site.company_name }}</a>
-        {%- endif %}
-    </div>
-    <div class="footer__github">
-        <span class="page__icon page__icon_github footer__github-icon"></span>
-        <a href="https://github.com/werf/werf" class="footer__github-counter">
-            <span class="footer__github-counter-num"><span class="gh_counter">1470</span></span>
-        </a>
-    </div>
-    <div class="footer__gen" id="siteGenerationTime" hidden>
-        {%- if page.lang == "ru" %}
-        Дата генерации сайта: {{ site.time | date: "%e.%m.%Y, %H:%M %z"  }}
-        {%- else %}
-        Site last generated: {{ site.time | date: "%b %-d, %Y at %H:%M %z"  }}
-        {%- endif %}
-    </div>
-    <div id="werfVersion" hidden>{{ site.werfVersion }}</div>
-    <div id="werfChannel" hidden>{{ site.werfChannel }}</div>
-</div>
