+++ docs/_includes/head.html
@@ -49,8 +49,4 @@
 <meta property="twitter:title" content="{{ generated_title }} | {{ site.site_title }}">
 <meta property="twitter:description" content="{{ description }}">
 
-{%- if site.ssi %}
 <!--#include virtual="/includes/head-common.html" -->
-{%- else %}
-{% include head-common.html %}
-{%- endif %}
