--- docs/_includes/ru/banner.html
@@ -1,1 +1,0 @@
-{%- include ru/guides_banner.html %}
