+++ docs/_layouts/sidebar-nosidebar.html
@@ -1,11 +1,7 @@
 ---
 layout: default
 ---
-{%- if page.lang == 'ru' %}
-{%- include ru/guides_banner.html %}
-{%- else %}
-{%- include en/guides_banner.html %}
-{%- endif %}
+<!--#include virtual="/includes/banner.html" -->
 <div class="page__container">
   {%- include nav-breadcrumbs.html %}
   <div class="layout-sidebar">
