+++ docs/_layouts/sidebar.html
@@ -1,11 +1,7 @@
 ---
 layout: default
 ---
-{%- if site.ssi %}
 <!--#include virtual="/includes/banner.html" -->
-{%- else %}
-{%- include {{ page.lang | append: "/banner.html" }} %}
-{%- endif %}
 <div class="page__container">
   {%- include nav-breadcrumbs.html %}
   <div class="layout-sidebar">
