--- docs/includes/banner.html
@@ -1,5 +1,0 @@
----
-layout: none
----
-
-{%- include {{ page.lang | append: "/banner.html" }} %}
