--- docs/includes/footer-common.html
@@ -1,5 +1,0 @@
----
-layout: none
----
-
-{% include footer-common.html %}
