+++ docs/pages/banner.html
@@ -1,0 +1,6 @@
+---
+layout: none
+permalink: /includes/banner.html
+---
+
+{%- include en/guides_banner.html %}
