+++ docs/pages/footer.html
@@ -1,0 +1,5 @@
+---
+layout: none
+permalink: /includes/footer.html
+---
+{% include footer.html %}
