+++ docs/pages/head-common.html
@@ -1,5 +1,5 @@
 ---
 layout: none
+permalink: /includes/head-common.html
 ---
-
 {% include head-common.html %}
