+++ docs/pages/index.md
@@ -2,6 +2,7 @@
 title: GitOps CLI tool
 permalink: /
 layout: default
+relation: site
 ---
 
 <div class="intro-scheme" id="intro-scheme">
