+++ docs/pages/publications.html
@@ -3,6 +3,7 @@ title: Publications
 permalink: publications.html
 layout: default
 description: News and articles related to werf
+relation: site
 ---
 
 {% asset publications.css %}
