+++ docs/pages/topnav.html
@@ -1,5 +1,5 @@
 ---
 layout: none
+permalink: /includes/topnav.html
 ---
-
 {% include topnav.html %}
