+++ docs/pages_ru/banner.html
@@ -1,0 +1,6 @@
+---
+layout: none
+permalink: /includes/banner.html
+---
+
+{%- include ru/guides_banner.html %}
