+++ docs/pages_ru/head-common.html
@@ -1,0 +1,5 @@
+---
+layout: none
+permalink: /includes/head-common.html
+---
+{% include head-common.html %}
