+++ docs/pages_ru/publications.html
@@ -2,6 +2,7 @@
 title: Публикации
 permalink: publications.html
 layout: default
+relation: site
 ---
 
 {% asset publications.css %}
