+++ docs/pages_ru/robots_ru.txt
@@ -1,3 +1,7 @@
+---
+layout: none
+permalink: /robots.txt
+---
 User-agent: *
 
 Disallow: /latest
