+++ docs/pages_ru/topnav.html
@@ -1,0 +1,5 @@
+---
+layout: none
+permalink: /includes/topnav.html
+---
+{% include topnav.html %}
