+++ docs/version-menu.html
@@ -1,0 +1,5 @@
+---
+layout: none
+permalink: /includes/version-menu.html
+---
+{% include version-menu.html %}
